fixed the signal handling
With systemd, a simple service restart would log a failure. It would
simply fail and exit with a 143 exit status (128 + SIGTERM).

   Stopping Hosts file updater daemon...
   hostsd.service: main process exited, code=exited, status=143/n/a
   Unit hostsd.service entered failed state.

Another problem was that the call to `inotifywait(1)` was blocking, and
the trap would not be called until something happened in HOSTSD_DIR. A
call to `systemctl stop hostsd` would forcefully kill the process, but
not report it as "failed".

The last issue was again with the call to `inotifywait(1)`. In case of a
non-zero exit status, the script would prematurely abort due to the use
of `set -e` at the beginning of the script.

To fix this, the trap has been split in two. Before the exit, we clean
the environment as usual. When SIGTERM is received, we explicitly exit
with a zero exit status. And to make sure the trap will be executed as
soon as the SIGTERM signal is received, `inotifywatch(1)` is run in the
background and immediately waited upon. And finally to avoid a spurious
exit, the `watch_files` function call is part of a boolean expression.